kill some warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@9278 c90b9560-bf6c-de11-be94-00142212c4b1

+++ main/php_ini.c
@@ -344,7 +344,7 @@ PHPAPI void display_ini_entries(zend_module_entry *module)
 	}
 	PUTS("<table border=5 width=\"600\">\n");
 	php_info_print_table_header(3, "Directive", "Local Value", "Master Value");
-	zend_hash_apply_with_argument(&known_directives, (int (*)(void *, void *)) php_ini_displayer, (void *) module_number);
+	zend_hash_apply_with_argument(&known_directives, (int (*)(void *, void *)) php_ini_displayer, (void *) (long) module_number);
 	PUTS("</table>\n");
 }
